Refactored code so that CRUD of student is admin's route & added a middleware to provide acess
