+++ app/Http/Controllers/API/StudentController.php
@@ -38,17 +38,6 @@ class StudentController extends Controller
 
 
 
-    public function getStudents()
-    {
-
-        $data = Student::all();
-
-        if ($data->isEmpty()) {
-            return jsonResponse(null, 'No students found in Database ', 404);
-        }
-
-        return jsonResponse($data, 'Students sucessfully retrieved', 200);
-    }
 
     /**
      * Store a newly created resource in storage.
@@ -84,76 +73,16 @@ class StudentController extends Controller
     /**
      * Display the specified resource.
      */
-    public function showStudent(Request $request, $student_id)
-    {
-        $data = Student::find($student_id);
-
-        if (!$data) {
-            return jsonResponse(null, 'Student was not found', 404);
-        }
-        return jsonResponse($data, 'Student was found', 200);
-    }
+  
 
     /**
      * Update the specified resource in storage.
      */
-    public function updateStudent(Request $request, $student_id)
-    {
-        Log::info($request->all());
-        $data = Student::find($student_id);
-
-        if (!$data) {
-            return jsonResponse(null, 'Student was not found', 404);
-        }
-
-
-        if ($request->hasFile('image')) {
-            if ($data->image_path) {
-                $filePath = public_path('images/' . $data->image_path);
-                if (file_exists($filePath)) {
-                    unlink($filePath);
-                }
-            }
-
-            // Upload new image
-            $newImageName = time() . '-' . $request->name . '.' . $request->image->extension();
-            $request->image->move(public_path('images'), $newImageName);
-
-            $data->image_path = $newImageName;
-        }
-
-        // Update other fields
-        $data->update([
-            'name' => $request->name,
-            'age' => $request->age,
-            'gender' => $request->gender,
-            'course' => $request->course,
-            'image_path' => $data->image_path ?? null
-        ]);
-
-        return jsonResponse($data, 'Student has been updated successfully', 201);
-    }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function deleteStudent($student_id)
-    {
-        $data = Student::find($student_id);
-        if (!$data) {
-            return jsonResponse(null, 'Student was not found', 404);
-        }
-
-        if ($data->image_path) {
-            $filePath = public_path('images/' . $data->image_path);
-            if (file_exists($filePath)) {
-                unlink($filePath);
-            }
-        }
-
-        Student::destroy($student_id);
-        return jsonResponse($data, 'Student has been deleted', 200);
-    }
+ 
 
     public function rentBook(Request $request, $student_id)
     {
